backend/app/requirements.txt
Flask
python-dotenv
boto3
flask-cors
pdf-lib
puppeteer